Add the handlers for login/out processes

+++ src/model/ticketsModel.js
@@ -1,10 +1,10 @@
 const pool = require('./pool');
-function getTicketFromDb(id, callback) {
-    console.log("Getting ticket from DB");
+function getTicketFromDb(data, callback) {
+    console.log("Getting ticket: " + data.ticket_id + ", from user: " + data.user_id + " on the DB");
   
-    const sql = "SELECT u.ticket_id, u.ticket_name, u.ticket_email, c.company_name  FROM tickets AS u LEFT JOIN companies AS c ON u.company_id=c.company_id WHERE u.ticket_id = $1::int";
+    const sql = "SELECT ticket_id, ticket_title, ticket_content FROM tickets WHERE ticket_id = $1::int AND user_id = $2::int";
   
-    const params = [id];
+    const params = [data.ticket_id, data.user_id];
   
     pool.query(sql, params, (err, result) => {
       if (err) {
@@ -12,31 +12,30 @@ function getTicketFromDb(id, callback) {
         console.log(err);
         callback(err, null);
       }
-      console.log(result.rows);
+      console.log('Sending ticket info to client');
       callback(null, result.rows);
     })
   };
   
-  function getTicketsByUserFromDb(id, callback) {
-    console.log("Getting tickets by user from DB");
+  function getTicketsByUserFromDb(user_id, callback) {
+    console.log("Getting tickets from user: " + user_id + " from DB");
   
     const sql = "SELECT ticket_date, ticket_id, ticket_title FROM tickets WHERE user_id = $1::int ORDER BY ticket_date desc";
-    let params = [id];
+    let params = [user_id];
 
     pool.query(sql, params, (err, result) => {
       if (err) {
         console.log("Error in query: ");
-        console.log(err);
         callback(err, null);
       } else {
-        console.log(result.rows);
+        console.log('Sending tickets summary to client');
         callback(null, result.rows);
       }
     })
   };
 
   function insertTicketToDb(data, callback) {
-    console.log("Inserting ticket to DB");
+    console.log("User " + data.user_id + " started the process to insert a new ticket on the DB");
   
     const sql = "INSERT INTO tickets (user_id, ticket_title, ticket_content) VALUES ($1, $2, $3)";
     
@@ -45,31 +44,29 @@ function getTicketFromDb(id, callback) {
     pool.query(sql, params, (err, result) => {
       if (err) {
         console.log("Error in query: ");
-        console.log(err);
         callback(err, null);
       } else {
         let myArray = [{recordsInserted: result.rowCount}];
-        console.log(myArray);
+        console.log(myArray[0].recordsInserted + ' new ticket was inserted to the DB');
         callback(null, myArray);
       }
     })
   };
 
-  function removeticketFromDb(id, callback) {
+  function removeticketFromDb(data, callback) {
     console.log("Removing ticket from DB");
   
-    const sql = "DELETE FROM tickets WHERE ticket_id = $1::int";
+    const sql = "DELETE FROM tickets WHERE ticket_id = $1::int and user_id = $2::int";
   
-    const params = [id];
+    const params = [data.ticket_id, data.user_id];
 
     pool.query(sql, params, (err, result) => {
       if (err) {
         console.log("Error in query: ");
-        console.log(err);
         callback(err, null);
       } else {
         let myArray = [{recordsDeleted: result.rowCount}];
-        console.log(myArray);
+        console.log(myArray[0].recordsDeleted + ' records were deleted from the DB');
         callback(null, myArray);
       }
 
@@ -79,18 +76,17 @@ function getTicketFromDb(id, callback) {
   function updateticketFromDb (data, callback) {
     console.log("Updating ticket from DB");
   
-    const sql = "UPDATE tickets SET ticket_name = $2, ticket_email = $3, ticket_password = $4 WHERE ticket_id = $1";
+    const sql = "UPDATE tickets SET ticket_name = $1, ticket_email = $2, ticket_password = $3 WHERE ticket_id = $4";
   
-    const params = [data.ticket_id, data.ticket_name, data.ticket_email, data.ticket_password];
+    const params = [data.ticket_name, data.ticket_email, data.ticket_password, data.ticket_id];
   
     pool.query(sql, params, (err, result) => {
       if (err) {
         console.log("Error in query: ");
-        console.log(err);
         callback(err, null);
       } else  {
         let myArray = [{recordsUpdated: result.rowCount}];
-        console.log(myArray);
+        console.log(myArray[0].recordsUpdated + ' records where updated on the DB');
         callback(null, myArray);
       }
   })
